# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CareerPathr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students/{userId}/profile:
    get:
      operationId: getStudentProfile
      tags: [students]
      summary: Get student profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateStudentProfile
      tags: [students]
      summary: Update student profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── ASSESSMENT ──────────────────────────────────────────────────────────
...
  /recommendations/{userId}:
    get:
      operationId: getUserRecommendations
      tags: [recommendations]
      summary: Get career recommendations for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of recommendations with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /recommendations/{userId}/skill-gap:
    get:
      operationId: getSkillGapAnalysis
      tags: [recommendations]
      summary: Get skill gap analysis for a user vs their top recommended career
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skill gap analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGapAnalysis"
        "404":
          description: No data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── COURSES ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StudentProfile:
      type: object
      required: [id, userId, yearOfStudy, major, skills, interests, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        yearOfStudy:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
    StudentProfileInput:
      type: object
      required: [yearOfStudy, major, skills, interests]
      properties:
        yearOfStudy:
          type: string
        major:
          type: string
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string

    # ─── ASSESSMENT ─────────────────────────────────────────────────────────
...
    Career:
      type: object
      required: [id, title, category, description, requiredSkills, avgSalary, growthRate, learningTimeMonths]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        avgSalary:
          type: string
        growthRate:
          type: string
        learningTimeMonths:
          type: integer
...
    Recommendation:
      type: object
      required: [id, userId, careerId, matchScore, career]
      properties:
        id:
          type: integer
        userId:
          type: integer
        careerId:
          type: integer
        matchScore:
          type: number
        career:
          $ref: "#/components/schemas/Career"
    SkillGapAnalysis:
      type: object
      required: [careerId, careerTitle, requiredSkills, userSkills, missingSkills, matchedSkills, matchPercentage]
      properties:
        careerId:
          type: integer
        careerTitle:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        userSkills:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        matchedSkills:
          type: array
          items:
            type: string
        matchPercentage:
          type: number

    # ─── COURSE ─────────────────────────────────────────────────────────────
...