# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CareerPathr API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Career } from './career';

export interface Recommendation {
  id: number;
  userId: number;
  careerId: number;
  matchScore: number;
  career: Career;
}