# lib/api-zod/src/generated/types/studentProfileInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CareerPathr API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfileInput {
  yearOfStudy: string;
  major: string;
  skills: string[];
  interests: string[];
}